Dashboard page: refactored setup of fetch mock, added more assertions to unit test
