+++ __mocks__/gatsby-plugin-firebase/auth/auth.js
@@ -2,7 +2,7 @@
 import { v4 as uuidv4 } from 'uuid';
 
 import Constants from '../constants/auth';
-import delay from '../../utils/index';
+import { delay } from '../../../src/utils/index';
 
 const singleton = Symbol('');
 const singletonEnforcer = Symbol('');
