--- __mocks__/utils/index.js
@@ -1,5 +1,0 @@
-const delay = async (milliseconds) => {
-  await new Promise((resolve) => setTimeout(resolve, milliseconds));
-};
-
-export default delay;
